==> AQUATOX/Code/R_Analyses/02_Model_Calibration_Effects/01_Compile_Effects_Data.Rmd <==
---
title: "Compile control data"
author: "A. Lindborg"
date: "5/5/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Purpose
Organize data from the effects studies into one master dataset. Data are provided by GmbH.

Output = Dataframe with the following columns
  * Study: 19b
  * Parameter Group: (e.g. chemical parameters, macrophytes, phytoplankton, etc.)
  * Units or ID: if a species will be units, if a chemical/phy parameter will be study ID
  * Date: POSIXct date, all in the same format
  * Enclosure: numbered 1-5 (this is one less than the cailbration data)
  * Variable: Species name or chemical/physical parameter
  * Value: measurement (biomass, count, etc.)


# History
5/5/2022 - A. Lindborg modifies script based on L. Mudge validation control script of same name


# Setup
```{r setup}
library(tidyverse)
library(openxlsx)
library(lubridate)

```

# Data
* Original Files from Waterborne are organized by experiment & functional group (or physical parameter)
* Excel file name components are separated by an underscore: experiment _ functional group _ experiment ID


# Read in Mesocosm Control Data

```{r read_files}

# Create a path for each experiment folder (note: not using the parent folder bc we don't want the files from _read-me or _received):

path19 <- "Data/Effects Data GmbH/Full Study 19b - AZT Study"


# Create a list of file names:
files <- list.files(path =  path19,
                    pattern = "*.csv",
                    full.names = TRUE)


# Create a df that has a nested list of the species data:
control_data_list <- tibble(filename = files) %>% 
  mutate(contents = map(filename, ~read_csv(.x, col_names = FALSE, skip_empty_rows = TRUE)), 
         file_details = map_chr(contents, ~.x[[1,1]]), #first row in each file has context for the data
         header_line = map(contents, ~.x[2, ]), # pull out the header line with col names
         data_contents = map2(contents, header_line, ~slice(.x, -c(1:2)) %>% set_names(.y))) %>%
  select(-contents, -header_line, -filename) %>%
  separate(col= file_details, into = c("study", "parameter_group", "units_or_id"), sep = "_") %>% #splits file details into 3 columns
  mutate(units_or_id = 
           case_when(parameter_group == "Emerging-Insects" ~ "[Individuals/Sample]", #manually edit formatting some units
                     parameter_group == "Phytoplankton-Chl-a" ~ "ug/L",
                     parameter_group == "Periphyton-Chl-a" ~ "ug/a",
                     TRUE ~ units_or_id),
         data_contents = map(data_contents, ~janitor::remove_empty(.x)), #renaming to match other study naming conventions
         study = ifelse(study == "Full-Study-19b", "Control-19b", study)) 

# Send individual data frames to the environment- named by parameter group
list2env(split(x=control_data_list, f=control_data_list$parameter_group), envir=.GlobalEnv) 

```

# Clean & Arrange Master Dataset

The below chunk creates the master dataset (includes all critters, all experiments) using the individual dataframes in the environment. The `Water-Level` and `Sediment-Water-Parameters` are not included in this data cleaning step because they have a different format than the rest of the data (and we don't need these right away)

```{r master_df}

# get a list of data frames- but first remove ones we don't need first (rm= remove)
#rm(dw)
rm(control_data_list)
rm(`Sediment-Water-Parameters`)
rm(`Water-Level`)
df_list <- Filter(function(x) is(x, "data.frame"), mget(ls()))


# Iterate over a list of data frames:
  # unnest the data_contents
  # pivot_longer
  # filter out any NA from value (so we aren't keeping empty rows)
  # remove < symbol from values
  # convert "value" to numeric class
  # Fix Date column- put into standard format
  # bind all data together

out <- c()

for(i in seq(1:length(df_list))){
  tmp <- df_list[[i]] %>%
    unnest(cols=c(data_contents)) %>%
    pivot_longer(cols=!c(study:Enclosure), names_to = "Variable", values_to = "Value") %>%
    filter(!is.na(Value)) %>%
    mutate(Value = as.numeric(gsub("<", "", Value)),
           Date = lubridate::parse_date_time(x=Date, orders = c("m/d/Y", "d/m/Y","m/d/y", "d.m.Y"), tz = "UTC"),
           date2= Date) # created 2ndary date column so can set all years to be the same, below
  
  out <- bind_rows(out, tmp)
  
  
}

lubridate::year(out$date2) <- 2021
out <- janitor::clean_names(out)

write_csv(out, "Output/01_Compile_Effects_Data/effects_long.csv")
saveRDS(out, file = "Output/01_Compile_Effects_Data/effects_long.rds")


# Try removing the unicode replacement symbol (black triangle with question mark)
#out$Variable <- enc2utf8(out$Variable)
#out$Variable <- gsub("/uFFFD", "", out$Variable, fixed=TRUE)
#out$Variable <- gsub("/u00B0", "", out$Variable, fixed=TRUE)

```

``` {r check species}

 cal <- read.csv("Working_Files/R_Analyses/00_Model_Calibration/Output/01_Compile_Control_Data/control_long.csv") 
 cal <- cal %>% 
   filter(study == "Control-19b") %>% #only compare to 19b control (not other years)
   select(variable)
   

 val <- out %>% 
   select(variable) 

 check = setdiff(cal$variable, val$variable)  #no differences between 2019 control and effects data

``` 


 # Subset data with calibration criteria 
 The purpose of this section is to create a subset output data file that only contains data for the groups that have calibration criteria- this is so that for future plots, etc. we don't have to repeat the filtering. 

 Calibration Documents from Waterborne (see folder by same name in R_analyses) 
 * `Criteria_ModelCalib_4Mar2021.docx` -- summary of all consensus groups 
 * `criteria_groups_M_Z_03Mar2021.xlsx` -- zooplankton and macroinvertebrates methods (incl. how individual species grouped) 
 * `criteria_groups_phyto_peri_3Mar2021.xlsx` -- phytoplankton and periphyton methods 
 * `ControlGraphs_2021-01-28.docx` -- ref that explains which life stages are added together for a total abundance per date 
 
``` {r calib_data}


 # Calibration criteria -- DO NOT MOVE EARLIER IN SCRIPT, will mess up creation of "out" 
 crit <- read.xlsx("Working_Files/R_Analyses/compiled_calibration_crit.xlsx", sheet = "calib_crit", detectDates = TRUE) %>% 
   rename(crit_min = min_value, 
          crit_max = max_value) 

 # Add Macrophytes to the crit dataset- not part of the official calibration criteria, but important for modeled groups! 
 macro <- out %>% filter(parameter_group == "Macrophytes") %>% distinct(parameter_group, variable, units_or_id) %>% rename(units=units_or_id) 
 crit <- bind_rows(crit, macro) 

 # Subset data 
 dsub <- semi_join(out, crit, by = c("parameter_group", "variable")) 

 #Check that the groups/species match EXACTLY between dsub and criteria: 
 ct <- crit %>% distinct(parameter_group, variable) %>% arrange(parameter_group, variable) 
 dt <- dsub %>% distinct(parameter_group, variable) %>% arrange(parameter_group, variable) 

 dplyr::all_equal(ct, dt) #TRUE 
 rm(ct) 
 rm(dt) 

 # Add consensus group 
 #dc <- left_join(dsub, crit %>% select(parameter_group, consensus_group, variable), by = c("parameter_group", "variable")) 
 dc <- left_join(dsub, crit, by = c("parameter_group", "variable")) %>% 
   relocate(enclosure, .after = study) %>% 
   relocate(c(consensus_group, variable), .after = parameter_group) %>% 
   relocate(date2, .after = date) %>% 
   dplyr::select(-units) #redundant 

# write_csv(dc, "Output/01_Compile_Effects_Data/effects_data_for_calibration.csv") 
# saveRDS(dc, "Output/01_Compile_Effects_Data/effects_data_for_calibration.rds") 

``` 
 
``` {r control_long_with_criteria}

 # Add calibration criteria to full control dataset 
 out.crit <- left_join(out, crit, by = c("parameter_group", "variable")) %>% 
   relocate(enclosure, .after = study) %>% 
   relocate(c(consensus_group, variable), .after = parameter_group) %>% 
   relocate(date2, .after = date) %>% 
   dplyr::select(-units) 


 #write_csv(out.crit, "Output/01_Compile_Effects_Data/effects_long_wCriteria.csv") 
 #saveRDS(out.crit, "Output/01_Compile_Effects_Data/effects_long_wCriteria.rds")   

```
